+++ src/affine/wide_float.rs
@@ -35,16 +35,8 @@ macro_rules! impl_wide_float {
                 Self::from_submatrix_translation(submatrix, translation)
             }
 
-            /// Returns the inverse of `self` or `None` if for any lane `self`
-            /// is not invertable.
-            #[inline]
-            #[must_use]
-            pub fn try_inverse(&self) -> Option<Self> {
-                let submatrix = self.submatrix.try_inverse()?;
-                let translation = -self.translation * submatrix;
-
-                Some(Self::from_submatrix_translation(submatrix, translation))
-            }
+            // `try_inverse` is exluded on purpose. It would not be useful
+            // because it would only return `Some` if all lanes succeed.
 
             /// For each lane, returns the inverse of `self` or `fallback` if
             /// `self` is not invertable.
@@ -407,9 +399,7 @@ mod tests {
 
     use crate::{
         Affine, Affine2, Affine3, EulerRot, Mat3, Quat, Unaligned, Vec2, Vec3, Vector,
-        utils::{
-            assert_panic_test_eq, assert_test_eq, assert_test_eq_or_panic, for_types, random_iter,
-        },
+        utils::{assert_test_eq, assert_test_eq_or_panic, for_types, random_iter},
     };
 
     #[test]
@@ -448,17 +438,7 @@ mod tests {
         });
     }
 
-    #[test]
-    fn test_try_inverse() {
-        for_types!(|N, Wide: WideFloat| {
-            for affine in random_iter::<Affine<N, Wide, Unaligned>>() {
-                assert_panic_test_eq!(
-                    affine.try_inverse().unwrap(),
-                    Affine::from_lane_fn(|lane| affine.lane(lane).try_inverse().unwrap())
-                );
-            }
-        });
-    }
+    // `try_inverse` is exluded on purpose.
 
     #[test]
     fn test_inverse_or() {